requirements.txt
# Flask Framework
Flask==3.0.0

# Database
mysql-connector-python==8.2.0

# Authentication & Security
bcrypt==4.1.2
pycryptodome==3.19.1
cryptography==42.0.5

# Email validation
email-validator==2.1.0.post1

# Environment variables (for config)
python-dotenv==1.0.1

# MSN: Task scheduler for automatic permission cleanup
APScheduler==3.10.4

# Additional utilities
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.5
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# Serial communication for NFC configurator
pyserial==3.5

flask_cors

flask_mysqldb